Rename folder preprocessing to object_seperating
Change the folder name according to the purpose of use
