+++ object_separating/std_dict.pickle
